ecommerce_backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
python-dotenv>=1.0
drf-yasg==1.21.11
graphene-django>=3.1
celery[redis]>=5.3
redis>=4.5
django-filter>=23.1
gunicorn>=20.1
